Feature/add current maxvel to awapi (#1142)
* add current max vel to awapi

* add readme

* apply format

+++ awapi/awapi_awiv_adapter/include/awapi_awiv_adapter/awapi_max_velocity_publisher.hpp
@@ -35,7 +35,7 @@ private:
   bool calcMaxVelocity(
     const autoware_api_msgs::msg::VelocityLimit::ConstSharedPtr & max_velocity_ptr,
     const autoware_api_msgs::msg::StopCommand::ConstSharedPtr & temporary_stop_ptr,
-    float * max_velocity);
+    double * max_velocity);
 
   double default_max_velocity_;
 };
